rename python package to librascal
