+++ bindings/librascal/__init__.py
@@ -1,3 +1,3 @@
 __all__ = ["lib", "representations", "utils", "neighbourlist"]
 
-__version__ = "0.0.0"
+__version__ = "0.1.0"
