+++ bindings/librascal/lib/__init__.py
