+++ bindings/librascal/models/sparse_points.py
