+++ bindings/librascal/neighbourlist/structure_manager.py
