+++ bindings/librascal/representations/coulomb_matrix.py
