+++ bindings/librascal/representations/spherical_expansion.py
