+++ bindings/librascal/utils/filter.py
