+++ bindings/librascal/utils/pool_worker.py
