+++ requirements/common.txt
@@ -1,0 +1,6 @@
+numpy>=1.16.0
+scipy>=1.4.0
+ase>=3.19.0
+sympy>=1.4
+# notebook tutorials
+matplotlib>=2.0.0
