+++ requirements/dev.txt
@@ -1,11 +1,7 @@
+-r common.txt
 # python library
-numpy
-scipy
 jupyter
-sympy>=1.4
 skcosmo
-# notebook tutorials
-matplotlib
 # Developers tools for compiling documentation
 sphinx>=2.1.2,<=4.0.2
 breathe>=4.14.1
