--- requirements_pip.txt
@@ -1,5 +1,0 @@
-# python library that contain numpy and scipy already
-numpy>=1.16.0
-scipy>=1.4.0
-matplotlib>=2.0.0
-ase>=3.19.0
