+++ tests/python/python_models_test.py
@@ -1,8 +1,8 @@
-from rascal.representations import SphericalInvariants
-from rascal.models import Kernel
-from rascal.models.sparse_points import SparsePoints
-from rascal.models.kernels import compute_numerical_kernel_gradients
-from rascal.utils import from_dict, to_dict
+from librascal.representations import SphericalInvariants
+from librascal.models import Kernel
+from librascal.models.sparse_points import SparsePoints
+from librascal.models.kernels import compute_numerical_kernel_gradients
+from librascal.utils import from_dict, to_dict
 from test_utils import load_json_frame, BoxList, Box, compute_relative_error
 from ase.calculators.lj import LennardJones
 import ase.io
